+++ app/src/main/kotlin/lgfs/gfs/ChunkMetadata.kt
@@ -1,3 +1,3 @@
 package lgfs.gfs
 
-data class ChunkMetadata(val handle: Long, val index : Int)
+data class ChunkMetadata(val handle: Long, val index: Long)
